Isolated ranks command in separate file, refactoring ranks command
